Toolbar.tsx: Abstract out toolbar as separate component
